[feat] move component to folder
